Move regenerateIds function to utils
