+++ apps/design/backend/src/app.ts
@@ -31,7 +31,6 @@ import {
   AnyContestSchema,
   HmpbBallotPaperSizeSchema,
   SystemSettingsSchema,
-  hasSplits,
 } from '@votingworks/types';
 import express, { Application } from 'express';
 import {
@@ -41,7 +40,6 @@ import {
   groupBy,
   ok,
   Result,
-  throwIllegalValue,
 } from '@votingworks/basics';
 import {
   BallotLayoutError,
@@ -78,7 +76,7 @@ import {
   authEnabled,
 } from './globals';
 import { createBallotPropsForTemplate, defaultBallotTemplate } from './ballots';
-import { generateId, getPdfFileName } from './utils';
+import { getPdfFileName, regenerateElectionIds } from './utils';
 import {
   ElectionFeaturesConfig,
   getElectionFeaturesConfig,
@@ -146,90 +144,6 @@ export function convertVxfPrecincts(election: Election): SplittablePrecinct[] {
       })),
     };
   });
-}
-
-function regenerateIds(
-  election: Election,
-  precincts: SplittablePrecinct[]
-): {
-  districts: District[];
-  precincts: SplittablePrecinct[];
-  parties: Party[];
-  contests: AnyContest[];
-} {
-  const idMap = new Map<string, string>();
-  function replaceId<T extends string>(id: T): T {
-    if (!idMap.has(id)) {
-      idMap.set(id, generateId());
-    }
-    return assertDefined(idMap.get(id)) as T;
-  }
-
-  const districts = election.districts.map((district) => ({
-    ...district,
-    id: replaceId(district.id),
-  }));
-  const updatedPrecincts = precincts.map((precinct) => {
-    if (hasSplits(precinct)) {
-      return {
-        ...precinct,
-        id: replaceId(precinct.id),
-        splits: precinct.splits.map((split) => ({
-          ...split,
-          id: replaceId(split.id),
-          districtIds: split.districtIds.map(replaceId),
-        })),
-      };
-    }
-    return {
-      ...precinct,
-      id: replaceId(precinct.id),
-      districtIds: precinct.districtIds.map(replaceId),
-    };
-  });
-  const parties = election.parties.map((party) => ({
-    ...party,
-    id: replaceId(party.id),
-  }));
-  const contests = election.contests.map((contest) => ({
-    ...contest,
-    id: replaceId(contest.id),
-    districtId: replaceId(contest.districtId),
-    ...(() => {
-      switch (contest.type) {
-        case 'candidate':
-          return {
-            partyId: contest.partyId ? replaceId(contest.partyId) : undefined,
-            candidates: contest.candidates.map((candidate) => ({
-              ...candidate,
-              id: replaceId(candidate.id),
-              partyIds: candidate.partyIds?.map(replaceId),
-            })),
-          };
-        case 'yesno':
-          return {
-            yesOption: {
-              ...contest.yesOption,
-              id: replaceId(contest.yesOption.id),
-            },
-            noOption: {
-              ...contest.noOption,
-              id: replaceId(contest.noOption.id),
-            },
-          };
-        default: {
-          /* istanbul ignore next - @preserve */
-          throwIllegalValue(contest);
-        }
-      }
-    })(),
-  }));
-  return {
-    districts,
-    precincts: updatedPrecincts,
-    parties,
-    contests,
-  };
 }
 
 const TextInput = z
@@ -334,7 +248,7 @@ function buildApi({ auth, workspace, translator }: AppContext) {
       if (parseResult.isErr()) return parseResult;
       const sourceElection = parseResult.ok();
       const sourcePrecincts = convertVxfPrecincts(sourceElection);
-      const { districts, precincts, parties, contests } = regenerateIds(
+      const { districts, precincts, parties, contests } = regenerateElectionIds(
         sourceElection,
         sourcePrecincts
       );
@@ -413,7 +327,7 @@ function buildApi({ auth, workspace, translator }: AppContext) {
         });
       }
 
-      const { districts, precincts, parties, contests } = regenerateIds(
+      const { districts, precincts, parties, contests } = regenerateElectionIds(
         sourceElection,
         sourcePrecincts
       );
